# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/space": {
            "get": {
                "tags": [
                    "Space"
                ],
                "summary": "List all enabled spaces",
                "operationId": "d86141499ab1e32ed7c3d0c1964184e7",
                "responses": {
                    "200": {
                        "description": "Spaces",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Small saloon"
                                            },
                                            "type": {
                                                "type": "string",
                                                "example": "saloon"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "A cozy and intimate space for small gatherings and events."
                                            },
                                            "capacity": {
                                                "type": "integer",
                                                "example": 10
                                            },
                                            "enable": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "image_url": {
                                                "type": "string",
                                                "example": "http://example.com/small_space.jpg"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Space"
                ],
                "summary": "Create a space",
                "operationId": "4d91b42a570a9811ab8eb557987c49bd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "description",
                                    "type",
                                    "capacity",
                                    "image"
                                ],
                                "properties": {
                                    "description": {
                                        "type": "string",
                                        "example": "A cozy and intimate space for small gatherings and events."
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "saloon"
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Result message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/space/{id}/daily-available-slots": {
            "get": {
                "tags": [
                    "Space"
                ],
                "summary": "List Available Time Slots for a Space",
                "operationId": "d4d39a798f8c53300ef0778bc3e027b8",
                "responses": {
                    "200": {
                        "description": "Available times",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "start": {
                                            "type": "string",
                                            "example": "08:00"
                                        },
                                        "end": {
                                            "type": "string",
                                            "example": "08:30"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/space/{id}": {
            "get": {
                "tags": [
                    "Space"
                ],
                "summary": "space",
                "operationId": "5e5c4bf62fc2942a880bb9e98e7c7136",
                "responses": {
                    "200": {
                        "description": "Space",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Small saloon"
                                        },
                                        "type": {
                                            "type": "string",
                                            "example": "saloon"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "A cozy and intimate space for small gatherings and events."
                                        },
                                        "capacity": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "enable": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "image_url": {
                                            "type": "string",
                                            "example": "http://example.com/small_space.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Space"
                ],
                "summary": "Update a space",
                "operationId": "61d610e4626177b0cfd827becd15ec57",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "description",
                                    "type",
                                    "capacity",
                                    "image"
                                ],
                                "properties": {
                                    "description": {
                                        "type": "string",
                                        "example": "A cozy and intimate space for small gatherings and events."
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "saloon"
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Result message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Space"
                ],
                "summary": "Delete reservation",
                "operationId": "9a02f09477ab5ae7069f5f545a6c6ab0",
                "responses": {
                    "200": {
                        "description": "Result messge",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}